Auditor API SDK: SHA Pinning workflow file (#3)
SHA pinning is a way to declare that a certain package is only
accessible for a given commit in the package.

This is to prevent attacks -- for example, a malicious actor could
poison Github and push a bad commit. SHA pinning will pin up to a
version that I've vetted with security.

We have updated our allow list to enable the SHA pinned version but not
the version tagged one so this PR updated that. Once this PR is merged,
then I can start generating packages.